requirements.txt
flask==0.12.2
flask_restful==0.3.6
flask_script==2.0.6
flask_migrate==2.1.1
gunicorn==20.1.0
python-dotenv==0.17.1
pymongo==3.11.4
dictionaries==0.0.2
dnspython==2.1.0
